+++ src/lib/components/document/edit-page/PaintingSVG.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState, MouseEvent } from 'react'
+import { useRef, useState } from 'react'
 import { useDocumentContext } from '../../hooks/useDocumentContext'
 import { EDIT_TOOLS } from '../../../../utils/enums'
 import useCustomEffect from '../../hooks/useCustomEffect'
@@ -12,8 +12,6 @@ const PaintingSVG = () => {
     activeEditTool,
     svgHeight,
     svgWidth,
-    elements,
-    setElements,
     editOpacity,
     layer,
   } = useDocumentContext()
@@ -32,7 +30,6 @@ const PaintingSVG = () => {
     entity: SVGLineElement | SVGRectElement | SVGPathElement
   ) => {
     if (EDIT_TOOLS.ERASER === activeEditTool && isDrawing) {
-      setElements(elements.filter((item) => entity !== item))
       svgRef.current.removeChild(entity)
     }
   }
@@ -61,11 +58,11 @@ const PaintingSVG = () => {
 
   // reactivate hover function in elements
   useCustomEffect(() => {
-    elements.forEach((element) => {
-      if (element) {
-        element.onmouseover = () => handleMouseOver(element)
+    for (const child of svgRef.current.children) {
+      if (child) {
+        child.onmouseover = () => handleMouseOver(child)
       }
-    })
+    }
   }, [isDrawing, activeEditTool])
 
   // Responsivness
@@ -73,67 +70,72 @@ const PaintingSVG = () => {
     const svg = svgRef.current
     const width = svg.width.baseVal.value
     const height = svg.height.baseVal.value
+    const children = svgRef.current.children
 
     console.log(width, height)
 
-    if (elements.length !== 0) {
-      elements.forEach((element) => {
-        // adaptd stroke width
-        if (element) {
-          const strokeWidth = element.getAttribute('stroke-width')
+    if (children.length !== 0) {
+      for (const child of children) {
+        // Adjust stroke width
+        if (child) {
+          const strokeWidth = child.getAttribute('stroke-width')
           if (strokeWidth) {
-            element.setAttribute(
-              'stroke-width',
-              Math.max(1, (parseInt(strokeWidth) / width) * svgWidth).toString()
+            const newStrokeWidth = (parseInt(strokeWidth) / width) * svgWidth
+            const clampedStrokeWidth = Math.max(1, newStrokeWidth) // Ensure a minimum stroke width
+            child.setAttribute('stroke-width', clampedStrokeWidth.toString())
+          }
+
+          // FOR STRAIGHT LINE
+          if (child instanceof SVGLineElement) {
+            const x1 = child.x1.baseVal.value
+            const y1 = child.y1.baseVal.value
+            const x2 = child.x2.baseVal.value
+            const y2 = child.y2.baseVal.value
+
+            // Adjust height and width for position
+            child.setAttribute('x1', ((x1 / width) * svgWidth).toString())
+            child.setAttribute('y1', ((y1 / height) * svgHeight).toString())
+            child.setAttribute('x2', ((x2 / width) * svgWidth).toString())
+            child.setAttribute('y2', ((y2 / height) * svgHeight).toString())
+          }
+          // FOR RECT
+          else if (child instanceof SVGRectElement) {
+            const x = child.x.baseVal.value
+            const y = child.y.baseVal.value
+            const entWidth = child.width.baseVal.value
+            const entHeight = child.height.baseVal.value
+
+            // Adjust height and width for position
+            child.setAttribute('x', ((x / width) * svgWidth).toString())
+            child.setAttribute('y', ((y / height) * svgHeight).toString())
+            child.setAttribute(
+              'width',
+              ((entWidth / width) * svgWidth).toString()
+            )
+            child.setAttribute(
+              'height',
+              ((entHeight / height) * svgHeight).toString()
             )
           }
-        }
-        // FOR STRAIGHT LINE
-        if (element && element instanceof SVGLineElement) {
-          const x1 = element.x1.baseVal.value
-          const y1 = element.y1.baseVal.value
-          const x2 = element.x2.baseVal.value
-          const y2 = element.y2.baseVal.value
+          // FOR LINE
+          else if (child instanceof SVGPathElement) {
+            const d = child.getAttribute('d') as string
+            if (d) {
+              const coords = d.split(' ').map((coord) => {
+                const command = coord[0]
+                const points = coord.substring(1).split(',')
+                const x = ((parseInt(points[0]) / width) * svgWidth).toFixed(0)
+                const y = ((parseInt(points[1]) / height) * svgHeight).toFixed(
+                  0
+                )
 
-          element.setAttribute('x1', ((x1 / width) * svgWidth).toString())
-          element.setAttribute('y1', ((y1 / height) * svgHeight).toString())
-          element.setAttribute('x2', ((x2 / width) * svgWidth).toString())
-          element.setAttribute('y2', ((y2 / height) * svgHeight).toString())
-        }
-        // FOR RECT
-        else if (element && element instanceof SVGRectElement) {
-          const x = element.x.baseVal.value
-          const y = element.y.baseVal.value
-          const entWidth = element.width.baseVal.value
-          const entHeight = element.height.baseVal.value
-
-          element.setAttribute('x', ((x / width) * svgWidth).toString())
-          element.setAttribute('y', ((y / height) * svgHeight).toString())
-          element.setAttribute(
-            'width',
-            ((entWidth / width) * svgWidth).toString()
-          )
-          element.setAttribute(
-            'height',
-            ((entHeight / height) * svgHeight).toString()
-          )
-        }
-        // FOR LINE
-        else if (element && element instanceof SVGPathElement) {
-          const d = element.getAttribute('d') as string
-          if (d) {
-            const coords = d.split(' ').map((coord) => {
-              const command = coord[0]
-              const points = coord.substring(1).split(',')
-              const x = ((parseInt(points[0]) / width) * svgWidth).toFixed(0)
-              const y = ((parseInt(points[1]) / height) * svgHeight).toFixed(0)
-
-              if (command && x && y) return `${command}${x},${y}`
-            })
-            element.setAttribute('d', coords.join(' '))
+                if (command && x && y) return `${command}${x},${y}`
+              })
+              child.setAttribute('d', coords.join(' '))
+            }
           }
         }
-      })
+      }
     }
 
     svg.setAttribute('width', svgWidth)
@@ -202,14 +204,14 @@ const PaintingSVG = () => {
       setIsDrawing(false)
       return
     }
-    if (isDrawing) elements.push(entity)
     setEntity(null)
     setIsDrawing(false)
     setPath('M')
     setRecreate((prev) => !prev) // for trigger createing new entity
   }
 
-  useEffect(() => {
+  // Copy svg layer to pating layer
+  useCustomEffect(() => {
     if (layer && svgRef.current) {
       // Clear existing SVG content
       svgRef.current.innerHTML = ''
@@ -229,21 +231,7 @@ const PaintingSVG = () => {
         )[] = []
         Array.from(parsedSvg.children).forEach((child) => {
           svgRef.current.appendChild(child)
-          switch (child.tagName.toLowerCase()) {
-            case 'line':
-              tmpElements.push(child as SVGLineElement)
-              break
-            case 'rect':
-              tmpElements.push(child as SVGRectElement)
-              break
-            case 'path':
-              tmpElements.push(child as SVGPathElement)
-              break
-            default:
-              break
-          }
         })
-        setElements(tmpElements)
       }
     }
   }, [layer])